requirements.txt
asgiref==3.7.2
Django==3.2
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.4
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
